+++ src/main/java/coordinatecalculator/CoordinateApplication.java
@@ -3,7 +3,7 @@ package coordinatecalculator;
 import coordinatecalculator.domain.Shape;
 import coordinatecalculator.factory.ShapeMaker;
 import coordinatecalculator.util.OutputManager;
-import coordinatecalculator.view.OutputView;
+import coordinatecalculator.util.OutputView;
 
 public class CoordinateApplication {
     public static void main(String[] args) {
