+++ src/main/java/coordinatecalculator/util/OutputManager.java
@@ -1,7 +1,6 @@
 package coordinatecalculator.util;
 
 import coordinatecalculator.domain.Shape;
-import coordinatecalculator.view.OutputView;
 import coordinatecalculator.visitor.ShapePrintVisitor;
 
 public class OutputManager {
